+++ test/helpers/application_helper_test.rb
@@ -15,4 +15,20 @@ class ApplicationHelperTest < ActionView::TestCase
   test 'github_repo_url points at the project repository' do
     assert_equal 'https://github.com/romkey/pdxhackerspace-prusa-print-history', github_repo_url
   end
+
+  test 'footer defaults use the app name' do
+    assert_equal '3D Printer History', app_name
+    assert_match(/^3D Printer History v/, footer_text)
+  end
+
+  test 'printer_idle_dot_class reflects PrusaLink reachability' do
+    printer = printers(:prusa_xl)
+    printer.update!(prusalink_key: 'secret', prusalink_reachable: true)
+
+    assert_equal 'status-success', printer_idle_dot_class(printer)
+
+    printer.update!(prusalink_reachable: false)
+
+    assert_equal 'status-danger', printer_idle_dot_class(printer)
+  end
 end